Rename threat_model.py as enforcer.py.
