+++ mart/configs/attack/enforcer/batch.yaml
@@ -1,4 +1,4 @@
 defaults:
   - constraints: null
 
-_target_: mart.attack.threat_model.BatchEnforcer
+_target_: mart.attack.enforcer.BatchEnforcer
