+++ mart/configs/attack/enforcer/default.yaml
@@ -1,4 +1,4 @@
 defaults:
   - constraints: null
 
-_target_: mart.attack.threat_model.Enforcer
+_target_: mart.attack.enforcer.Enforcer
